Refactor of simulator, solver. Add option for selective sim.
