+++ src/attpc_engine/detector/solver.py
@@ -1,12 +1,19 @@
 import math
 import numpy as np
+from numpy.random import Generator
+from scipy.integrate import solve_ivp
 
 from spyral_utils.nuclear.target import GasTarget
 from spyral_utils.nuclear import NucleusData
 
+from .parameters import Config, DetectorParams
+from .typed_dict import NumbaTypedDict
+from .transporter import transport_track
 from .constants import MEV_2_JOULE, MEV_2_KG, C, E_CHARGE
 
 KE_LIMIT = 1e-6  # 1 eV
+# Time steps to solve ODE at. Each step is 1e-10 s
+TIME_STEPS = np.linspace(0, 10e-7, 10001)
 
 
 def equation_of_motion(
@@ -231,3 +238,157 @@ def rho_bound_condition(
 
     """
     return float(np.linalg.norm(state[:2])) - 0.292
+
+
+def generate_trajectory(
+    vertex: np.ndarray,
+    momentum: np.ndarray,
+    nucleus: NucleusData,
+    params: DetectorParams,
+) -> np.ndarray:
+    """Solves EoM for a nucleus in the AT-TPC
+
+    Solution is evaluated over a fixed time interval.
+
+    Parameters
+    ----------
+    config: Config
+        The simulation configuration
+
+    Returns
+    -------
+    numpy.ndarray
+        Nx6 array where each row is a solution to one of the ODEs evaluated at
+        the Nth time step.
+    """
+    # Find initial state of nucleus. (x, y, z, px, py, pz)
+    initial_state: np.ndarray = np.zeros(6)
+    initial_state[:3] = vertex  # m
+    initial_state[3:] = momentum[:3] / nucleus.mass  # unitless (gamma * beta)
+
+    # Set ODE stop conditions. See SciPy solve_ivp docs
+    stop_condition.terminal = True
+    stop_condition.direction = -1.0
+    forward_z_bound_condition.terminal = True
+    forward_z_bound_condition.direction = 1.0
+    backward_z_bound_condition.terminal = True
+    backward_z_bound_condition.direction = -1.0
+    rho_bound_condition.terminal = True
+    rho_bound_condition.direction = 1.0
+
+    track = solve_ivp(
+        equation_of_motion,
+        (0.0, 1.0),  # Set upper time limit to 1 sec. This should never be reached
+        initial_state,
+        method="Radau",
+        events=[
+            stop_condition,
+            forward_z_bound_condition,
+            backward_z_bound_condition,
+            rho_bound_condition,
+        ],
+        t_eval=TIME_STEPS,
+        args=(
+            params.bfield * -1.0,
+            params.efield * -1.0,
+            params.gas_target,
+            nucleus,
+        ),
+    )
+
+    return track.y.T  # Return transpose to easily index by row
+
+
+def generate_electrons(
+    track: np.ndarray, nucleus: NucleusData, params: DetectorParams, rng: Generator
+):
+    """Find the number of electrons made at each point of the nucleus' track.
+
+    Parameters
+    ----------
+    config: Config
+        The simulation configuration
+    track: numpy.ndarray
+        Nx6 array where each row is a solution to one of the ODEs evaluated at
+        the Nth time step.
+    rng: numpy.random.Generator
+        numpy random number generator
+
+    Returns
+    -------
+    numpy.ndarray
+        Returns an array of the number of electrons made at each
+        point in the trajectory
+    """
+    # Find energy of nucleus at each point of its track
+    gv = np.linalg.norm(track[:, 3:], axis=1)
+    beta = np.sqrt(gv**2.0 / (1.0 + gv**2.0))
+    gamma = gv / beta
+    energy = nucleus.mass * (gamma - 1.0)  # MeV
+
+    # Find number of electrons created at each point of its track
+    electrons = np.zeros_like(energy)
+    electrons[1:] = abs(np.diff(energy))  # Magnitude of energy lost
+    electrons *= 1.0e6 / params.w_value  # Convert to eV to match units of W-value
+
+    # Adjust number of electrons by Fano factor, can only have integer amount of electrons
+    electrons = np.array(
+        [rng.normal(point, np.sqrt(params.fano_factor * point)) for point in electrons],
+        dtype=np.int64,
+    )
+    return electrons
+
+
+def generate_point_cloud(
+    momentum: np.ndarray,
+    vertex: np.ndarray,
+    nucleus: NucleusData,
+    config: Config,
+    rng: Generator,
+    points: NumbaTypedDict,
+) -> None:
+    """Create the point cloud
+
+    Finds the pads hit by the electrons transported from each point
+    of the nucleus' trajectory to the pad plane.
+
+    Parameters
+    ----------
+    config: Config
+        The simulation configuration
+    rng: numpy.random.Generator
+        numpy random number generator
+    points: numba.typed.Dict[int, int]
+        A dictionary mapping a unique pad,tb key to the number of electrons.
+    """
+    # Generate nucleus' track and calculate the electrons made at each point
+    track = generate_trajectory(vertex, momentum, nucleus, config.det_params)
+    electrons = generate_electrons(track, nucleus, config.det_params, rng)
+
+    # Remove points in trajectory that create less than 1 electron
+    mask = electrons >= 1
+    track = track[mask]
+    electrons = electrons[mask]
+
+    # Apply gain factor from micropattern gas detectors
+    electrons *= config.det_params.mpgd_gain
+
+    # Convert z position of trajectory to exact time buckets
+    dv = config.drift_velocity
+    track[:, 2] = (
+        config.det_params.length - track[:, 2]
+    ) / dv + config.elec_params.micromegas_edge
+
+    if config.pad_grid_edges is None or config.pad_grid is None:
+        raise ValueError("Pad grid is not loaded at SimParticle.generate_hits!")
+
+    transport_track(
+        config.pad_grid,
+        config.pad_grid_edges,
+        config.det_params.diffusion,
+        config.det_params.efield,
+        config.drift_velocity,
+        track,
+        electrons,
+        points,
+    )
